Fix device index resolution in DCP A2A workspace init
`mr.device` resolves to `"cuda"` (index=None) on this path, which made
`can_use_custom_all_reduce_v2`/NVLink P2P access checks evaluate against
the default device rather than the rank-local GPU. Pass
`cuda:{mr.gpu_id}` so the symmetric-memory A2A workspace init sees the
correct device.
